+++ .gitignore
@@ -1,60 +1,38 @@
-# compiled output
-/dist
-/node_modules
-/build
+# prod
+dist/
 
-# Logs
-logs
-*.log
-npm-debug.log*
-pnpm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-lerna-debug.log*
-
-# OS
-.DS_Store
-
-# Tests
-/coverage
-/.nyc_output
-
-# IDEs and editors
-/.idea
-.project
-.classpath
-.c9/
-*.launch
-.settings/
-*.sublime-workspace
-
-# IDE - VSCode
+# dev
+.yarn/
+!.yarn/releases
 .vscode/*
 !.vscode/settings.json
 !.vscode/tasks.json
 !.vscode/launch.json
 !.vscode/extensions.json
+.idea/workspace.xml
+.idea/usage.statistics.xml
+.idea/shelf
 
-# dotenv environment variable files
+# deps
+node_modules/
+.wrangler
+
+# env
 .env
-.env.development.local
-.env.test.local
-.env.production.local
-.env.local
+.env.production
+.dev.vars
 
-# temp directory
-.temp
-.tmp
+# logs
+logs/
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
 
-# Runtime data
-pids
-*.pid
-*.seed
-*.pid.lock
-
-# Diagnostic reports (https://nodejs.org/api/report.html)
-report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
+# misc
+.DS_Store
 
 test/mock
 test/creds.json
-.vercel
