+++ package.json
@@ -1,55 +1,30 @@
 {
   "name": "cloud-sync",
-  "version": "0.0.1",
-  "description": "",
-  "author": "",
-  "private": true,
-  "license": "MIT",
   "scripts": {
-    "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\"",
-    "start": "nest start ",
-    "start:dev": "nest start --watch",
-    "start:debug": "nest start --debug --watch",
-    "start:prod": "node dist/main",
-    "lint": "eslint \"src/**/*.ts\" --fix"
+    "lint": "eslint \"src/**/*.ts\" --fix",
+    "dev": "wrangler dev",
+    "deploy": "wrangler deploy --minify"
   },
   "dependencies": {
-    "@nestjs/common": "^10.3.9",
-    "@nestjs/config": "^3.2.2",
-    "@nestjs/core": "^10.3.9",
-    "@nestjs/platform-express": "^10.3.9",
     "@resvg/resvg-js": "^2.6.2",
-    "@upstash/redis": "^1.31.5",
-    "cloudflare": "^3.4.0",
-    "discord-api-types": "^0.37.90",
-    "joi": "^17.13.1",
-    "jose": "^5.4.0",
-    "lz-string": "^1.5.0",
-    "reflect-metadata": "^0.2.2",
-    "rxjs": "^7.8.1"
+    "cloudflare": "^3.5.0",
+    "discord-api-types": "^0.37.105",
+    "hono": "^4.6.10",
+    "joi": "^17.13.3",
+    "jose": "^5.9.6"
   },
   "devDependencies": {
-    "@nestjs/cli": "^10.3.2",
-    "@nestjs/schematics": "^10.1.1",
-    "@nestjs/testing": "^10.3.9",
-    "@types/express": "^4.17.21",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^20.14.2",
-    "@types/supertest": "^6.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.21.0",
-    "@typescript-eslint/parser": "^6.21.0",
-    "eslint": "^8.57.0",
+    "@cloudflare/workers-types": "^4.20241022.0",
+    "@types/node": "^22.9.0",
+    "@typescript-eslint/eslint-plugin": "^8.14.0",
+    "@typescript-eslint/parser": "^8.14.0",
+    "eslint": "^9.15.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "jest": "^29.7.0",
-    "prettier": "^3.3.2",
-    "source-map-support": "^0.5.21",
-    "supertest": "^6.3.4",
-    "ts-jest": "^29.1.4",
-    "ts-loader": "^9.5.1",
+    "eslint-plugin-prettier": "^5.2.1",
+    "prettier": "^3.3.3",
     "ts-node": "^10.9.2",
-    "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.4.5"
+    "typescript": "^5.6.3",
+    "wrangler": "^3.83.0"
   }
 }
