+++ src/index.ts
@@ -1,0 +1,7 @@
+import { Hono } from "hono";
+
+const app = new Hono();
+
+app.get("/", (c) => c.redirect("https://github.com/nexpid/CloudSync", 301));
+
+export default app;
