--- vercel.json
@@ -1,22 +1,0 @@
-{
-  "version": 2,
-  "builds": [
-    {
-      "src": "src/main.ts",
-      "use": "@vercel/node"
-    }
-  ],
-  "routes": [
-    {
-      "src": "/(.*)",
-      "dest": "src/main.ts",
-      "methods": ["GET", "POST", "PUT", "DELETE"]
-    }
-  ],
-  "crons": [
-    {
-      "path": "/api/silly",
-      "schedule": "0 2 * * *"
-    }
-  ]
-}
