+++ wrangler.toml
@@ -1,0 +1,31 @@
+name = "cloudsync"
+main = "src/index.ts"
+compatibility_date = "2024-11-16"
+
+compatibility_flags = [ "nodejs_compat_v2" ]
+
+[triggers]
+crons = ["0 2 * * *"]
+
+# [vars]
+# MY_VAR = "my-variable"
+
+# [[kv_namespaces]]
+# binding = "MY_KV_NAMESPACE"
+# id = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+
+# [[r2_buckets]]
+# binding = "MY_BUCKET"
+# bucket_name = "my-bucket"
+
+# [[d1_databases]]
+# binding = "DB"
+# database_name = "my-database"
+# database_id = ""
+
+# [ai]
+# binding = "AI"
+
+# [observability]
+# enabled = true
+# head_sampling_rate = 1
